Add last 2 pages
